+++ cd1d_importer.user.js
@@ -1,5 +1,5 @@
 // ==UserScript==
-// @name         Import CD1D releases to MusicBrainz (⚠️ deprecated, no-op)
+// @name         [DISCONTINUED] Import CD1D releases to MusicBrainz
 // @description  [⚠️ script is deprecated] CD1D importer is deprecated since CD1D.com no longer exists. This script is a no-op now and will be removed in the future. If for some reason you still have it installed, it is time to remove it.
 // @namespace    http://userscripts.org/users/517952
 // @match        https://*.musicbrainz.org/*
