+++ src/userscripts/albumlink_importer/meta.json
@@ -1,13 +1,13 @@
 {
     "name": "Import album.link releases to MusicBrainz",
-    "description": "Import album.link smart links with Harmony and add their remaining URL relationships to MusicBrainz.",
-    "version": "2026.09.12.1",
+    "description": "Deprecated: install the unified Smartlink importer, then remove this script.",
+    "version": "2026.09.13.1",
     "author": "Raman Sinclair",
     "namespace": "https://github.com/murdos/musicbrainz-userscripts/",
     "downloadURL": "https://raw.githubusercontent.com/murdos/musicbrainz-userscripts/dist/albumlink_importer.user.js",
     "updateURL": "https://raw.githubusercontent.com/murdos/musicbrainz-userscripts/dist/albumlink_importer.user.js",
     "match": ["https://album.link/*", "https://*.album.link/*"],
-    "grant": ["GM.getValue", "GM.setValue", "GM_getValue", "GM_setValue"],
+    "discontinued": true,
     "runAt": "document-idle",
     "icon": "https://metabrainz.org/static/img/projects/musicbrainz.svg"
 }
