+++ src/userscripts/bfan_importer/meta.json
@@ -1,13 +1,13 @@
 {
     "name": "Import bfan.link releases to MusicBrainz",
-    "description": "Import bfan.link smart links with Harmony and add their remaining URL relationships to MusicBrainz. Bfan is Believe Digital's link aggregator service.",
-    "version": "2026.09.08.1",
+    "description": "Deprecated: install the unified Smartlink importer, then remove this script.",
+    "version": "2026.09.13.1",
     "author": "Raman Sinclair",
     "namespace": "https://github.com/murdos/musicbrainz-userscripts/",
     "downloadURL": "https://raw.githubusercontent.com/murdos/musicbrainz-userscripts/dist/bfan_importer.user.js",
     "updateURL": "https://raw.githubusercontent.com/murdos/musicbrainz-userscripts/dist/bfan_importer.user.js",
     "match": ["https://bfan.link/*", "https://*.bfan.link/*"],
-    "grant": ["GM.getValue", "GM.setValue", "GM_getValue", "GM_setValue"],
+    "discontinued": true,
     "runAt": "document-idle",
     "icon": "https://metabrainz.org/static/img/projects/musicbrainz.svg"
 }
