+++ src/userscripts/smartlink_importer/utils/extractors/bfan.ts
@@ -1,17 +1,15 @@
-import { normalizeServiceName } from '~/lib/smart-link-importer/logic';
+import { normalizeServiceName } from '~/userscripts/smartlink_importer/utils/logic';
+import { record } from '~/userscripts/smartlink_importer/utils/misc/record';
+import type { ServiceElement } from '~/userscripts/smartlink_importer/utils/types';
 
-export interface BfanServiceData {
+interface BfanServiceData {
     service: string;
     label: string;
     action: string;
     sourceUrl: string;
 }
 
-function record(value: unknown): Record<string, unknown> | undefined {
-    return value && typeof value === 'object' && !Array.isArray(value) ? (value as Record<string, unknown>) : undefined;
-}
-
-/** Read release-provider destinations from bfan.link's Next.js hydration payload. */
+/** Read release-provider destinations from bfan.link’s Next.js hydration payload. */
 export function extractBfanServiceData(payload: unknown): BfanServiceData[] {
     const root = record(payload);
     const props = record(root?.['props']);
@@ -45,4 +43,34 @@ export function extractBfanServiceData(payload: unknown): BfanServiceData[] {
         });
     }
     return links;
+}
+
+function readServiceData(): BfanServiceData[] {
+    const nextData = document.querySelector<HTMLScriptElement>('script#__NEXT_DATA__')?.textContent;
+    if (!nextData) return [];
+    try {
+        return extractBfanServiceData(JSON.parse(nextData) as unknown);
+    } catch {
+        return [];
+    }
+}
+
+export function collectBfanServiceElements(): ServiceElement[] {
+    const dataByService = new Map(readServiceData().map(data => [data.service, data]));
+    const elements: ServiceElement[] = [];
+    for (const element of document.querySelectorAll<HTMLElement>('[data-testid="call-to-actions"] > [data-testid]')) {
+        const rawService = element.dataset['testid'] ?? '';
+        const service = normalizeServiceName(rawService);
+        const data = dataByService.get(service);
+        if (!data) continue;
+        elements.push({
+            cacheKey: service,
+            element,
+            service,
+            label: element.querySelector<HTMLImageElement>('img[alt]')?.alt || data.label,
+            action: element.querySelector<HTMLButtonElement>('button')?.textContent.trim() || data.action,
+            sourceUrl: data.sourceUrl,
+        });
+    }
+    return elements;
 }
