+++ src/userscripts/smartlink_importer/utils/extractors/common.ts
@@ -1,0 +1,5 @@
+export function nextCacheKey(counters: Map<string, number>, service: string): string {
+    const count = (counters.get(service) ?? 0) + 1;
+    counters.set(service, count);
+    return count === 1 ? service : `${service}:${count}`;
+}
