+++ src/userscripts/smartlink_importer/utils/misc/record.ts
@@ -1,0 +1,4 @@
+/** Return an object record when the value is a non-array object. */
+export function record(value: unknown): Record<string, unknown> | undefined {
+    return value && typeof value === 'object' && !Array.isArray(value) ? (value as Record<string, unknown>) : undefined;
+}
